Updates for 1121 release for Windows
Make the 1121 release build on Windows.  This is sufficient for getting
a basic Hello World dispatch program to run under swift again.

+++ src/semaphore_internal.h
@@ -32,7 +32,7 @@ struct dispatch_queue_s;
 DISPATCH_CLASS_DECL(semaphore, OBJECT);
 struct dispatch_semaphore_s {
 	DISPATCH_OBJECT_HEADER(semaphore);
-	long volatile dsema_value;
+	intptr_t volatile dsema_value;
 	intptr_t dsema_orig;
 	_dispatch_sema4_t dsema_sema;
 };
